# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gis Backend API",
        "description": "Dokumentasi API untuk platform Gis Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exam/user-answers": {
            "post": {
                "tags": [
                    "UserAnswer"
                ],
                "summary": "Submit jawaban user",
                "operationId": "c5a84cc16f931f86f47ad43c2844b8a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "question_id",
                                    "answer_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "question_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "answer_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jawaban disimpan"
                    },
                    "422": {
                        "description": "Jawaban tidak sesuai dengan soal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exam/user-answers/{id}": {
            "delete": {
                "tags": [
                    "UserAnswer"
                ],
                "summary": "Batalkan jawaban user",
                "operationId": "76af2e2083c048bee4b683bfd4c5a957",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jawaban user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jawaban dibatalkan"
                    },
                    "404": {
                        "description": "Jawaban tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exam/user-answers/{user_id}": {
            "get": {
                "tags": [
                    "UserAnswer"
                ],
                "summary": "List jawaban user berdasarkan user_id",
                "operationId": "3bd2fe27dabd798db0bbd16150d14cfc",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar jawaban user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exam/user-answers/{id}/toggle-doubt": {
            "patch": {
                "tags": [
                    "UserAnswer"
                ],
                "summary": "Toggle status ragu-ragu pada jawaban",
                "operationId": "fe3a78a1a7e032b595e72ec514609af6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jawaban user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status ragu-ragu diperbarui"
                    },
                    "404": {
                        "description": "Jawaban tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exam/user-answers/{id}/unset-doubt": {
            "patch": {
                "tags": [
                    "UserAnswer"
                ],
                "summary": "Batalkan status ragu-ragu pada jawaban user (set is_doubtful ke false)",
                "operationId": "88b8c05a6e495e049951a3536d7dabb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jawaban user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status ragu-ragu dibatalkan"
                    },
                    "404": {
                        "description": "Jawaban tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exam/results/{user_id}": {
            "get": {
                "tags": [
                    "UserAnswer"
                ],
                "summary": "Hitung hasil jawaban user",
                "operationId": "e9de97ac6cad5b8532eda2a1fc8a5cd3",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ringkasan hasil user (benar, salah, kosong)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}